+++ src/main/java/org/voyanttools/trombone/tool/corpus/DocumentsMetadata.java
@@ -132,7 +132,7 @@ public class DocumentsMetadata extends AbstractCorpusTool {
 	}
 	
 	@Override
-	public int getVersion() {
+	public float getVersion() {
 		return super.getVersion()+9;
 	}
