+++ src/main/java/org/voyanttools/trombone/tool/utils/RunnableTool.java
@@ -31,6 +31,6 @@ public interface RunnableTool {
 
 	public void run() throws IOException;
 	
-	public int getVersion();
+	public float getVersion();
 
 }
